Adds Projects Page Structure and Styling
Migrates the projects view from the old site to
this project. Required many styling tweaks after
moving content over. Biggest differences between
rendering the old site and react site is that
.container is a bootstrap holdover class that
did more than was apparent on the old site,
and that default font-size (of a <p> for example)
varies greatly between JSX and plain HTML.